+++ lib/components/List/List.d.ts
@@ -1,7 +1,9 @@
-import { ReactNode } from 'react';
+import { ReactElement } from 'react';
+import { IListItemProps } from './ListItem';
 interface IListProps {
-    children?: ReactNode;
+    children?: ReactElement<IListItemProps> | Array<ReactElement<IListItemProps>>;
     isFlush?: boolean;
+    isHoverable?: boolean;
 }
-export declare const List: ({ children, isFlush }: IListProps) => JSX.Element;
+export declare const List: ({ children, isFlush, isHoverable }: IListProps) => JSX.Element;
 export {};