+++ lib/components/List/list.models.d.ts
@@ -1,0 +1,8 @@
+import { ReactElement } from 'react';
+import { IListItemProps } from './ListItem';
+export declare class ListItemModel {
+    constructor(dto: ReactElement<IListItemProps>);
+    key: any;
+    props: IListItemProps;
+    type: any;
+}