+++ src/lib/components/List/list.models.ts
@@ -1,0 +1,12 @@
+import { ReactElement } from 'react';
+import { IListItemProps } from './ListItem';
+
+export class ListItemModel {
+	constructor(dto: ReactElement<IListItemProps>) {
+		return {...dto}
+	}
+
+	key: any;
+	props: IListItemProps;
+	type: any;
+}